initial commit move harmony utils to parser folder
