+++ vllm/entrypoints/openai/parser/harmony_utils.py
